Update the landmines game to keep scores and statistics per server. The /stats/landmines page has been removed in favor of a per-guild stat page.
This allows anyone to use the game on his own server.

There are a few DB-related caveats in this update, due in most part to the data migration and a bug in tortoise, the ORM used in the bot itself (https://github.com/tortoise/tortoise-orm/issues/822)

The bug makes it so that we *can't* have member, a OneToOneField, be the LandminesUserData primary key because tortoise fails to generate correct SQL queries when running with a OneToOneField as a primary_key.

There will be an available optimisation if/when a fix for https://github.com/tortoise/tortoise-orm/issues/822 gets released: remove the implicit id column and use member as the table primary key again.